requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-index --output-file=requirements/dev.txt requirements/base.txt requirements/dev.in requirements/testing.in
#
appdirs==1.4.3            # via black
attrs==19.3.0             # via black
beautifulsoup4==4.8.1     # via webtest
black==19.10b0            # via -r requirements/testing.in
certifi==2019.9.11        # via -r requirements/base.txt, requests
chardet==3.0.4            # via -r requirements/base.txt, requests
click==7.0                # via -r requirements/base.txt, black, pip-tools
coreapi==2.3.3            # via -r requirements/base.txt, drf-yasg
coreschema==0.0.4         # via -r requirements/base.txt, coreapi, drf-yasg
coverage==4.5.4           # via -r requirements/testing.in
django-better-admin-arrayfield==1.0.4  # via -r requirements/base.txt
django-choices==1.7.0     # via -r requirements/base.txt, vng-api-common
django-cors-middleware==1.4.0  # via -r requirements/base.txt
django-debug-toolbar==2.0  # via -r requirements/dev.in
django-extensions==2.2.3  # via -r requirements/dev.in
django-filter==2.2.0      # via -r requirements/base.txt, vng-api-common
django-markup==1.3        # via -r requirements/base.txt
django-redis==4.10.0      # via -r requirements/base.txt
django-rest-framework-condition==0.1.1  # via -r requirements/base.txt, vng-api-common
django-solo==1.1.3        # via -r requirements/base.txt, vng-api-common
django-webtest==1.9.7     # via -r requirements/testing.in
django==2.2.11            # via -r requirements/base.txt, django-choices, django-debug-toolbar, django-filter, django-markup, django-redis, django-rest-framework-condition, drf-nested-routers, drf-yasg, vng-api-common
djangorestframework-camel-case==1.1.1  # via -r requirements/base.txt, vng-api-common
djangorestframework==3.9.4  # via -r requirements/base.txt, drf-nested-routers, drf-yasg, vng-api-common
drf-nested-routers==0.91  # via -r requirements/base.txt, vng-api-common
drf-yasg==1.16.0          # via -r requirements/base.txt, vng-api-common
factory-boy==2.12.0       # via -r requirements/testing.in
faker==2.0.3              # via factory-boy
freezegun==0.3.12         # via -r requirements/testing.in
gemma-zds-client==0.13.0  # via -r requirements/base.txt, vng-api-common
idna==2.8                 # via -r requirements/base.txt, requests
inflection==0.3.1         # via -r requirements/base.txt, drf-yasg
iso-639==0.4.5            # via -r requirements/base.txt, vng-api-common
isodate==0.6.0            # via -r requirements/base.txt, vng-api-common
isort==4.3.21             # via -r requirements/testing.in
itypes==1.1.0             # via -r requirements/base.txt, coreapi
jinja2==2.10.3            # via -r requirements/base.txt, coreschema
markdown==3.1.1           # via -r requirements/base.txt
markupsafe==1.1.1         # via -r requirements/base.txt, jinja2
oyaml==0.9                # via -r requirements/base.txt, vng-api-common
pathspec==0.6.0           # via black
pip-tools==4.2.0          # via -r requirements/base.txt
psycopg2==2.8.3           # via -r requirements/base.txt
pyjwt==1.7.1              # via -r requirements/base.txt, gemma-zds-client, vng-api-common
python-dateutil==2.8.0    # via -r requirements/base.txt, faker, freezegun
python-dotenv==0.10.3     # via -r requirements/base.txt
pytz==2019.3              # via -r requirements/base.txt, django
pyyaml==5.1.2             # via -r requirements/base.txt, gemma-zds-client, oyaml, vng-api-common
raven==6.10.0             # via -r requirements/base.txt
redis==3.3.11             # via -r requirements/base.txt, django-redis
regex==2019.11.1          # via black
requests==2.22.0          # via -r requirements/base.txt, coreapi, gemma-zds-client, vng-api-common
ruamel.yaml.clib==0.2.0   # via -r requirements/base.txt, ruamel.yaml
ruamel.yaml==0.16.5       # via -r requirements/base.txt, drf-yasg
six==1.12.0               # via -r requirements/base.txt, django-extensions, django-markup, drf-yasg, faker, freezegun, isodate, pip-tools, python-dateutil, webtest
soupsieve==1.9.4          # via beautifulsoup4
sqlparse==0.3.0           # via -r requirements/base.txt, django, django-debug-toolbar
tblib==1.4.0              # via -r requirements/testing.in
text-unidecode==1.3       # via faker
toml==0.10.0              # via black
typed-ast==1.4.0          # via black
unidecode==1.1.1          # via -r requirements/base.txt, vng-api-common
uritemplate==3.0.0        # via -r requirements/base.txt, coreapi, drf-yasg
urllib3==1.25.6           # via -r requirements/base.txt, requests
uwsgi==2.0.18             # via -r requirements/base.txt
vng-api-common==1.4.0     # via -r requirements/base.txt
waitress==1.3.1           # via webtest
webob==1.8.5              # via webtest
webtest==2.0.33           # via django-webtest

# The following packages are considered to be unsafe in a requirements file:
# setuptools
